kern_resources_new/creative_lab/token_tracker/requirements.txt
litellm>=1.0.0
flask>=2.0.0
dash>=2.0.0
plotly>=5.0.0
pandas>=1.3.0
tiktoken>=0.5.0
python-dotenv>=0.19.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.7.0
gunicorn>=20.1.0
dash-bootstrap-components>=1.0.0
dash-auth>=1.4.0
email-validator>=2.0.0
PyJWT>=2.6.0
flask-cors>=3.0.10

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
